+++ app/src/main/AndroidManifest.xml
@@ -448,7 +448,7 @@
         <activity android:name="org.autojs.autojs.ui.settings.AboutActivity_" />
         <activity
             android:name="org.autojs.autojs.ui.settings.SettingsActivity"
-            android:theme="@style/AppTheme.Settings" />
+            android:theme="@style/M3AppTheme" />
         <activity android:name="org.autojs.autojs.ui.error.ErrorReportActivity" />
         <activity
             android:name="org.autojs.autojs.external.tasker.TaskerScriptEditActivity_"
@@ -576,14 +576,16 @@
 
         <activity
             android:name="org.autojs.autojs.external.open.RunIntentActivity"
+            android:exported="true"
+            android:enabled="false"
             android:icon="@drawable/autojs_logo"
             android:label="@string/text_run_script"
             android:taskAffinity=".external.open.RunIntentActivity"
             android:theme="@android:style/Theme.NoDisplay" />
         <activity
             android:name="org.autojs.autojs.external.open.OpenIntentActivity"
-            android:exported="true"
             android:excludeFromRecents="true"
+            android:exported="true"
             android:icon="@drawable/autojs_logo"
             android:label="@string/app_name"
             android:theme="@style/M3AppTheme.Transparent">
