+++ app/src/main/java/org/autojs/autojs/timing/work/AlarmManagerProvider.kt
@@ -13,7 +13,6 @@ import android.os.Build
 import android.os.SystemClock
 import android.util.Log
 import org.autojs.autojs.App
-import org.autojs.autoxjs.BuildConfig
 import org.autojs.autojs.external.ScriptIntents
 import org.autojs.autojs.timing.TimedTask
 import org.autojs.autojs.timing.TimedTaskManager
@@ -145,7 +144,7 @@ object AlarmManagerProvider : TimedTaskScheduler() {
                 Intent(ACTION_CHECK_TASK)
                     .setComponent(
                         ComponentName(
-                            BuildConfig.APPLICATION_ID,
+                            context.packageName,
                             "org.autojs.autojs.timing.work.AlarmManagerProvider"
                         )
                     ),
