+++ autojs/src/main/java/com/stardust/autojs/core/looper/Loopers.kt
@@ -169,13 +169,11 @@ class Loopers(val runtime: ScriptRuntime) {
         //mHandler.post(EMPTY_RUNNABLE)
     }
 
+    fun addAsyncTaskToCurrentThreadLooper(task: AsyncTask) {
+         (Thread.currentThread() as? TimerThread)?.loopers?.addAsyncTask(task) ?: addAsyncTask(task)
+    }
     companion object {
         private const val LOG_TAG = "Loopers"
         private val EMPTY_RUNNABLE = Runnable {}
-        fun addAsyncTaskToCurrentThreadLooper(task: AsyncTask): Boolean {
-            return (Thread.currentThread() as? TimerThread)?.loopers?.apply {
-                addAsyncTask(task)
-            } != null
-        }
     }
 }
