+++ autojs/src/main/java/com/stardust/autojs/runtime/ScriptBridges.kt
@@ -7,6 +7,7 @@ import org.mozilla.javascript.BaseFunction
 import org.mozilla.javascript.BoundFunction
 import org.mozilla.javascript.Context
 import org.mozilla.javascript.NativeJavaMethod
+import org.mozilla.javascript.NativeJavaObject
 import org.mozilla.javascript.Scriptable
 import org.mozilla.javascript.Undefined
 import org.mozilla.javascript.annotations.JSFunction
@@ -64,7 +65,7 @@ class ScriptBridges {
 
     fun asArray(obj: UiObjectCollection): Any = useJsContext { context ->
         val arr = toArray(obj.mNodes)
-        val thzs = Context.javaToJS(obj, arr) as Scriptable
+        val thzs = NativeJavaObject(arr, obj, obj.javaClass)
         obj::class.java.methods.forEach {
             val name = it.name
             val method = NativeJavaMethod(it, name)
