+++ automator/src/main/java/com/stardust/view/accessibility/NodeInfo.kt
@@ -4,19 +4,15 @@ import android.content.Context
 import android.content.pm.PackageManager
 import android.content.res.Resources
 import android.graphics.Rect
-import androidx.annotation.Keep
 import android.view.accessibility.AccessibilityNodeInfo
-
+import androidx.annotation.Keep
 import com.stardust.automator.UiObject
-
-import java.util.ArrayList
-import java.util.HashMap
 
 /**
  * Created by Stardust on 2017/3/10.
  */
 @Keep
-class NodeInfo(resources: Resources?, node: UiObject, var parent: NodeInfo?) {
+class NodeInfo(resources: Resources?,val node: UiObject, var parent: NodeInfo?) {
 
     private val children = ArrayList<NodeInfo>()
     val boundsInScreen = Rect()
